melhorias e correções (017-021)

+++ desafios/ex017.py
@@ -1,4 +1,4 @@
 from math import hypot
-oposto = int(input("Digite o comprimento do cateto oposto: "))
-adjacente = int(input("Digite o comprimento do cateto adjacente: "))
+oposto = float(input("Digite o comprimento do cateto oposto: "))
+adjacente = float(input("Digite o comprimento do cateto adjacente: "))
 print("\nO comprimento da hipotenusa é de {:.2f}".format(hypot(oposto, adjacente)))
